fix: correct commands and events to support postgresq

+++ src/commands/say.ts
@@ -8,11 +8,11 @@ export default class SayCommand extends BaseCommand {
   }
 
   async execute(client: Bot, message: Message, args: string[]): Promise<void> {
-    const channelId = args[0] && args[0].replace(/\D/g, "");
+    const channelId = args[1] && args[1].replace(/\D/g, "");
 
-    if (channelId) {
+    if (channelId && args[0]) {
       const channel = client.channels.cache.get(channelId) as TextChannel;
-      channel.send(args[1]);
+      channel.send(args[0]);
       message.delete();
 
       return;
